Set MAKEFLAGS to enable parallel make support
Cargo has a jobserver and passes it to build scripts by default,
but while the jobserver can act as a make-compatible *client* (e.g.
cargo build being called in a Makefile), it can't act as a
make-compatible *server*, not unless we pass on MAKEFLAGS
to sub-makes.

TLDR: previously, the build was single-threaded, now it's
multi threaded.