Use compile time detection of what features are available
Previously features were gated on the botan3 feature (which is removed here).
This didn't work well in practice as it made it difficult to add support for
features added in Botan3 after the 3.0 release.

Add declarations in botan-sys for functions added to ffi.h since 3.0.0, gated on
the detected library version. Also add interfaces for some of the new
functionality to the high level Rust crate.